requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
kubernetes==29.0.0
google-generativeai==0.8.2
prometheus-client==0.19.0
pydantic==2.5.0
httpx==0.26.0
python-multipart==0.0.6
python-dotenv==1.0.0